refactor: restructure hooks folder with domain based organization
